move StatementExecuteGroupEngine to jdbc package
